Update description component to match refined mocks. Refactor. #84.
